requirements.txt
qrcode
pillow
crc16
